moved components to correct folder
