+++ rosen_frontend/components/ProgressiveStepper/selectableImage.tsx
