+++ rosen_frontend/pages/train-algorithm.tsx
@@ -3,8 +3,8 @@ import ProgressiveStepper from "../components/ProgressiveStepper";
 import {Button, Step, StepLabel, Stepper, TextField} from "@mui/material";
 import {stepperTexts,stepperSteps} from "../utils/stepperText" 
 import { LinearProgress } from '@mui/material';
-import selectableImageGallery from './selectableImageGallery';
-import SelectableImage from './selectableImage';
+import selectableImageGallery from '../components/ProgressiveStepper/selectableImageGallery';
+import SelectableImage from '../components/ProgressiveStepper/selectableImage';
 
 
 const TrainAlgorithm = () => {
